# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: S9 AI Arena · API de plataforma
  # 0.2.0 · H6 (issue #10): añade getBotRatingHistory y getTeamStandings.
  # 0.3.0 · R8 admin ops: añade getSystemStatus y getRbacMatrix (solo lectura, admin).
  # 0.4.0 · R11 espectador público (slice mínimo): añade listPublicLiveBattles
  # (gateado por S9_PUBLIC_SPECTATE_ENABLED, apagado por defecto).
  # Añadir un endpoint = MINOR (docs/compatibilidad.md, §2 Semver por paquete).
  version: 0.4.0
  description: |
    Contrato de la API de plataforma (capítulos 16, 17 y 23 del Dosier técnico).

    **Autorización.** Cada operación declara `x-min-role` con el rol mínimo requerido.
    La jerarquía es acumulativa: visitor < user < developer < team_captain < organizer < moderator < admin.
    El middleware de E7 lee esta extensión: la interfaz web solo oculta, nunca autoriza.
    Además del rol, toda operación sobre un recurso propio exige **autorización de objeto**
    (ser dueño o miembro del equipo dueño); un rol suficiente no da acceso a recursos ajenos.

    **Privacidad.** Los campos marcados con `x-private: true` no aparecen NUNCA en respuestas
    a `visitor`. El test de fuga por contrato de E7/T7.5 recorre esta extensión.

    **Paginación** por cursor opaco (`cursor` + `limit`), nunca por offset.
servers:
  - url: https://arena.example/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /bots/{botId}/versions/{version}/actions/submit:
    post:
      tags: [bots]
      operationId: submitBotVersion
      summary: 'draft → validating. Lanza el pipeline de bot-manager (E6/T6.1).'
      x-min-role: developer
      parameters: [{ $ref: '#/components/parameters/BotId' }, { $ref: '#/components/parameters/Version' }]
      responses:
        '202': { description: Pipeline encolado, content: { application/json: { schema: { $ref: '#/components/schemas/Build' } } } }
        '409': { $ref: '#/components/responses/IllegalTransition' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BotId: { name: botId, in: path, required: true, schema: { type: string } }
...
    Version: { name: version, in: path, required: true, schema: { type: integer, minimum: 1 } }
  responses:
...
    IllegalTransition:
      description: 'Transición de estado ilegal (cap. 17.1). El cuerpo indica estado actual y transiciones permitidas.'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  currentState: { $ref: '#/components/schemas/BotState' }
                  allowedTransitions: { type: array, items: { type: string } }
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Código estable, apto para máquinas }
        message: { type: string, description: Explicación legible }
        correlationId: { type: string, description: Enlaza con los logs estructurados (cap. 24) }
...
    BotState:
      enum: [draft, validating, rejected, validated, published, frozen, suspended, retired]
      description: Estados del capítulo 17.1
...
    Build:
      type: object
      properties:
        id: { type: string }
        botId: { type: string }
        version: { type: integer }
        status: { enum: [queued, running, passed, failed] }
        stages:
          type: array
          description: Etapas del pipeline del capítulo 18.1, en orden
          items:
            type: object
            properties:
              name:
                enum: [structure, static_analysis, dependencies, build, protocol_test, smoke_battle, resource_limits, secret_scan, sign, publish]
              status: { enum: [pending, running, passed, failed, skipped] }
              message: { type: string }
              logUrl: { type: string, x-private: true, description: Solo dueño, moderador o admin }
        artifactHash: { type: string }
        createdAt: { type: string, format: date-time }
...